functions/requirements.txt
firebase-admin
firebase-functions
flask
google-generativeai
requests
opencv-python
numpy